Share the manifest test harness and split the suite by concern
